Final commit - Preview

+++ src/main/java/com/university/Services/supportServices/SupportService.java
@@ -5,7 +5,7 @@ import com.university.models.Statistic;
 
 public interface SupportService {
     public Integer countOfEmployee(Department department);
-    public Double avgSalary(Department department);
+    public Float avgSalary(Department department);
     public Statistic stat(Department department);
     public void changeUserStatus(Long userId,String status);
 }
